Test github action

+++ kita/Domain/Entities/PlaybackState.cs
@@ -4,8 +4,8 @@ namespace Kita.Domain
     {
         public string? CurrentSongId { get; set; }
         public bool IsPlaying { get; set; }
-        public int CurrentTime { get; set; } // in seconds
-        public int Volume { get; set; } = 100;
+        public double CurrentTime { get; set; }  // seconds (float from JS)
+        public double Volume { get; set; } = 100;
         public string? PlaylistId { get; set; }
         public List<string> Queue { get; set; } = new();
         public DateTime LastUpdated { get; set; } = DateTime.UtcNow;
